requirements.txt
streamlit==1.15.0
torch==2.0.0
transformers==4.30.0
Pillow==9.4.0
